requirements.txt
--extra-index-url https://test.pypi.org/simple/

# Core A2A and Server dependencies
a2a-sdk==0.2.11
a2a-sdk[sqlite]==0.2.11
uvicorn==0.35.0
httpx==0.28.1
fastapi==0.115.13
python-dotenv

# Semantic Kernel and Agent dependencies
semantic-kernel==1.34.0
microsoft-agents-core
microsoft-agents-copilotstudio-client
openai>=1.0.0

# Azure AI dependencies for MCP agents
azure-ai-projects
azure-ai-agents==1.1.0b4
azure-identity

# Additional utilities
rich
